Fix signedness warnings in gio/gdbusdaemon.c
gio/gdbusdaemon.c: In function ‘is_key’:
gio/gdbusdaemon.c:211:11: error: comparison of integer expressions of different signedness: ‘gsize’ {aka ‘long unsigned int’} and ‘long int’ [-Werror=sign-compare]
   if (len != key_end - key_start)
           ^~
gio/gdbusdaemon.c: In function ‘match_new’:
gio/gdbusdaemon.c:447:17: error: comparison of integer expressions of different signedness: ‘int’ and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
   for (i = 0; i < elements->len; i++)
                 ^

+++ gio/gdbusdaemon.c
@@ -208,7 +208,7 @@ is_key (const char *key_start, const char *key_end, const char *value)
 {
   gsize len = strlen (value);
 
-  if (len != key_end - key_start)
+  if (len != (gsize) (key_end - key_start))
     return FALSE;
 
   return strncmp (key_start, value, len) == 0;
@@ -358,7 +358,7 @@ match_new (const char *str)
   MatchElement element;
   gboolean eavesdrop;
   GDBusMessageType type;
-  int i;
+  guint i;
 
   eavesdrop = FALSE;
   type = G_DBUS_MESSAGE_TYPE_INVALID;
